[travis] Build lightning-net-tokio on Rust 1.39.0+, fuzz on stable
This disables building the lightning-net-tokio crate on 1.34.2 on
Travis in anticiption of relying on async/await for
lightning-net-tokio, with the nice side-effect of reducing the
amount of work done on our slowest Travis job.

This also adds 1.39.0 to Travis as that is the new MSRV for
lightning-net-tokio and uses that for codecov generation.

This also moves fuzzing to Rust stable, as honggfuzz broke its MSRV
and it seems likely it will again in the future.